+++ packages/arm-core/src/components/Scale.tsx
@@ -1,6 +1,6 @@
-import React, { useEffect } from 'react'
 import { useMapInstance } from '@datapunt/react-maps'
 import { Control, control } from 'leaflet'
+import React, { useEffect } from 'react'
 
 type Props = {
   options?: Control.ScaleOptions
@@ -9,17 +9,15 @@ type Props = {
 // Todo: Move this to react-maps
 const Scale: React.FC<Props> = ({ options }) => {
   const mapInstance = useMapInstance()
+
   useEffect(() => {
-    if (mapInstance) {
-      control.scale(options).addTo(mapInstance)
-    }
+    control.scale(options).addTo(mapInstance)
 
     return () => {
-      if (mapInstance) {
-        control.scale().remove()
-      }
+      control.scale().remove()
     }
-  }, [mapInstance])
+  }, [])
+
   return null
 }