+++ CHANGELOG.md
@@ -14,11 +14,15 @@ Prefix the change with one of these keywords:
 - Fixed: for any bug fixes.
 - Security: in case of vulnerabilities.
 
-## [0.4.0] (unreleased)
+## UNRELEASED
+
 - Added: `setInstance` prop to Marker component
 - Changed: MarkerClusterGroup has a breaking change. The L.MarkerClusterGroup layer is constructed outside of the component. 
 - Added: MarkerClusterGroup accepts events for the L.MarkerClusterGroup layer instead of individual markers
 - Added: MarkerClusterGroup now has a setInstance method
+- Changed: **BREAKING** `DrawTool` no longer accepts the `mapInstance` prop.
+- Changed: Return values of `useMapInstance` no longer checked for null values.
+- Changed: Updated version of `@datapunt/react-maps` to version 0.9.1
 
 ## [0.3.5]
 - Changed: Marker component now accepts the more generic LatLngExpression instead of only a LatLng